feat: pointer move/enter/exit events, drag widgets, coroutine Sleep, -Werror
Event dispatch:
- Add Widget::OnPointerMove/OnPointerEnter/OnPointerExit virtual hooks
- GlfwBridge: MouseMoveCallback + CursorPosCallback invokes it
- Application::DispatchPointerMove with DPI scaling and enter/exit detection
- HitTest cache with dirty invalidation on layout changes (avoids per-move
  full-tree traversal for large widget trees)

Coroutine Sleep:
- EventLoop timer queue (multimap<time_point, coroutine_handle>)
- ScheduleSleep() + thread-local Current() pointer
- SleepAwaitable::await_suspend registers with active loop
- RunReadyCoroutines resumes expired timers each frame

Drag & Scroll:
- Draggable widget: kIdle/kHovering/kDragging state machine, paint-time
  translate so Taitank layout is unaffected
- ScrollView: ScrollState (kIdle/kDragging) local to derived class,
  pointer-drag scrolling alongside wheel scrolling
- drag_demo example: Draggable box with state reporting + long-press
  coroutine demonstrating 'state machine as condition lock' pattern
  (weak_ptr + IsDragging() check acts as implicit cancellation)

Build:
- -Wall -Wextra -Wpedantic -Werror for neoflux target (0 warnings)
- -Wno-conversion/-Wno-sign-conversion for third-party/render code
- Fix member init order in RenderLayer
- --render_backend defaults to 'vulkan', adds 'cpu' option

Tests: 31/31 passing. clang-tidy: 0 warnings.

+++ src/render/tgfx_renderer.cpp
@@ -738,7 +738,8 @@ class GlRendererImpl : public NonCopyable {
     } else {
       // Negative pitch: bitmap is bottom-up. Read from bottom row.
       for (int row = 0; row < h; ++row) {
-        std::memcpy(packed.data() + static_cast<std::size_t>(row) * w,
+        std::memcpy(packed.data() + static_cast<std::size_t>(row) *
+                                        static_cast<std::size_t>(w),
                     bitmap.buffer + (h - 1 - row) * (-pitch),
                     static_cast<std::size_t>(w));
       }
